fix(ui/tests): resolve e2e failures, remove mock data, fix vitest config

+++ src/server/workers/proactive_analysis_job.rs
@@ -129,7 +129,7 @@ impl ProactiveAnalysisWorker {
                                 },
                                 _ => {
                                     attempts += 1;
-                                    tokio::time::sleep(std::time::Duration::from_secs(2u64.pow(attempts as u32))).await;
+                                    tokio::time::sleep(std::time::Duration::from_secs(2u64.pow(attempts))).await;
                                 }
                             }
                         }
